Moved file processor into a dedicated folder, and uploaded framework info to supa and chroma, made sure user uploads work
