+++ pyproject.toml
@@ -7,7 +7,13 @@ requires-python = ">=3.12"
 
 # Core dependencies shared by all services (if any)
 dependencies = [
+    "chromadb>=1.0.12",
+    "langchain-google-genai>=2.0.10",
+    "pypandoc>=1.15",
+    "pypdf2>=3.0.1",
     "python-dotenv>=1.1.0",
+    "reflex>=0.7.14",
+    "tiktoken>=0.9.0",
 ]
 
 [project.optional-dependencies]
@@ -43,6 +49,8 @@ dev = [
     "google-generativeai>=0.8.5",
     "pypandoc>=1.15",
     "pypdf2>=3.0.1",
+    "langchain-google-genai>=2.0.10",
+    "pydantic>=2.11.3",
 ]
 
 [dependency-groups]
